requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
matplotlib>=3.3.0
numpy>=1.20.0
streamlit>=1.10.0
streamlit-drawable-canvas>=0.8.0
pillow>=8.0.0
pycocotools>=2.0.0
onnx>=1.10.0
onnxruntime>=1.8.0
opencv-python-headless>=4.5.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.5
tqdm>=4.65.0